+++ generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/api/v36/converters/IssuesConverter.java
@@ -246,7 +246,7 @@ public class IssuesConverter {
                 metadataMap.put(metadata.getKey(), metadata);
         if (null != model.getIssues()) {
             for (IssueBase issue : model.getIssues()) {
-                List<BaseIssue> issues = convert(issue, metadataMap, dictionary);
+                List<BaseIssue> issues = convert(scanResult.getId(), issue, metadataMap, dictionary);
                 if (null != issues && !issues.isEmpty())
                     res.getIssues().addAll(issues);
                 else
@@ -325,6 +325,7 @@ public class IssuesConverter {
      * @param destination PT AI API version independent base issue
      */
     protected static void setBaseFields(
+            @NonNull final UUID scanResultId,
             @NonNull final IssueBase source,
             @NonNull final BaseIssue destination,
             @NonNull final Map<String, Map<Reports.Locale, ScanResult.Strings>> dictionary) {
@@ -336,7 +337,7 @@ public class IssuesConverter {
         destination.setFavorite(source.getIsFavorite());
         destination.setSuppressed(source.getIsSuppressed());
         destination.setSuspected(source.getIsSuspected());
-        destination.setNewInScanResultId(source.getIsNewInScanResultId());
+        destination.setIsNew(scanResultId.equals(source.getIsNewInScanResultId()));
 
         destination.setTypeId(source.getType());
     }
@@ -444,6 +445,7 @@ public class IssuesConverter {
      * @return PT AI API version independent vulnerability instance
      */
     protected static List<BaseIssue> convert(
+            @NonNull final UUID scanResultId,
             @NonNull final IssueBase issueBase,
             @NonNull final Map<String, IssueBaseMetadata> metadataMap,
             @NonNull final Map<String, Map<Reports.Locale, ScanResult.Strings>> dictionary) {
@@ -474,7 +476,7 @@ public class IssuesConverter {
                 V36BlackBoxIssue issue = (V36BlackBoxIssue) issueBase;
 
                 BlackBoxIssue res = new BlackBoxIssue();
-                setBaseFields(issue, res, dictionary);
+                setBaseFields(scanResultId, issue, res, dictionary);
                 applyMetadata(baseMetadata, res);
 
                 return Collections.singletonList(res);
@@ -482,7 +484,7 @@ public class IssuesConverter {
             V36ConfigurationIssue issue = (V36ConfigurationIssue) issueBase;
 
             ConfigurationIssue res = new ConfigurationIssue();
-            setBaseFields(issue, res, dictionary);
+            setBaseFields(scanResultId, issue, res, dictionary);
             applyMetadata(baseMetadata, res);
 
             res.setVulnerableExpression(convert(issue.getVulnerableExpression()));
@@ -514,7 +516,7 @@ public class IssuesConverter {
                 // type field and linked using fingerprintId instead. Let's init it with fingerprintId
                 // to avoid setBaseFields change
                 issue.setType(fingerprintId);
-                setBaseFields(issue, res, dictionary);
+                setBaseFields(scanResultId, issue, res, dictionary);
                 // As single V36FingerprintIssue may have multiple fingerprint IDs
                 // PT AI looks these IDs for maximum severity level and assigns it
                 // to issue. But as we decide to create individual ScaIssue for
@@ -544,13 +546,13 @@ public class IssuesConverter {
             V36UnknownIssue issue = (V36UnknownIssue) issueBase;
 
             UnknownIssue res = new UnknownIssue();
-            setBaseFields(issue, res, dictionary);
+            setBaseFields(scanResultId, issue, res, dictionary);
             return Collections.singletonList(res);
         } else if (IssueType.Vulnerability == issueType) {
             V36VulnerabilityIssue issue = (V36VulnerabilityIssue) issueBase;
 
             VulnerabilityIssue res = new VulnerabilityIssue();
-            setBaseFields(issue, res, dictionary);
+            setBaseFields(scanResultId, issue, res, dictionary);
             applyMetadata(baseMetadata, res);
 
             res.setSecondOrder(issue.getIsSecondOrder());
@@ -574,7 +576,7 @@ public class IssuesConverter {
             V36WeaknessIssue issue = (V36WeaknessIssue) issueBase;
 
             WeaknessIssue res = new WeaknessIssue();
-            setBaseFields(issue, res, dictionary);
+            setBaseFields(scanResultId, issue, res, dictionary);
             applyMetadata(baseMetadata, res);
             res.setVulnerableExpression(convert(issue.getVulnerableExpression()));
 
@@ -583,7 +585,7 @@ public class IssuesConverter {
             V36YaraMatchIssue issue = (V36YaraMatchIssue) issueBase;
 
             YaraMatchIssue res = new YaraMatchIssue();
-            setBaseFields(issue, res, dictionary);
+            setBaseFields(scanResultId, issue, res, dictionary);
             return Collections.singletonList(res);
         } else
             return null;